Some mailman installation have hidden inputs
The hidden field is setup to avoid replay...
So, we keep it (but not in the dump) and we use it at submit time.
Also handle the 'SUBMIT' type with the same mechanism is case of it would also be required

Note : hidden is in lower case... It can break if it is in upper case elsewhere.
The good solution would probably be to use xpath.